Community Page Preview and UI Optimizations
Added the Community Page for communities. It's still under development, but the main layout has been designed and implemented, along with its AJAX-like loading of posts.

Optimized the navbar area to darken the top of the website as soon as the user starts scrolling from the top of the page, so that the navbar is more visible when there's content behind it.

For UX purposes, the text editor toolbar is now sticky and follows the viewport as the user scrolls down on the edit post page.