Removed Chats Logic
Due to the scope of the project, the Chats will be removed for now.

All the chats and messages related code were causing trouble and conflicts with some of the other components.

The decision to eliminate the logic and functionality of the chats will allow the project to move forward more agilely, focusing on the most important aspects.